requirements.txt
strawberry-graphql==0.158.0
pytest==7.2.1